requirements.txt
# Core
huggingface_hub>=0.20.0
requests>=2.31.0
structlog>=24.1.0
python-dotenv>=1.0.0

# NLP
spacy>=3.7.0
openai>=1.12.0

# Data Processing
orjson>=3.9.0

# Semantic Clustering
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Database
pymongo>=4.6.0

# Groq
groq>=0.4.0
